requirements.txt
asgiref==3.2.7
Django==3.0.5
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
Pillow==9.1.1
django-environ==0.8.1
psycopg2==2.8.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6